public/memoverride: Fix aligned delete crashes on Windows in DEBUG mode
Operator new in DEBUG mode has special version which tracks file / line.
But for such case aligned delete is callled so we allocate memory as not
aligned but free as aligned.

It leads to crash as memory address to free computed incorrectly. As Windows
allocator uses 16 bytes alignment just check pointer alignment and if it is 16
than default allocator was used. And we can free memory by default method.