[filesystem] optimize file search by splitting it into two passes
Pass 1.
We first go thru all search paths and try to open it using CWin32ReadOnlyFile.
In most cases the file can be opened using it, so we can skip CStdioFile in this pass which should be a good speedup.

Pass 2.
We failed to open it using CWin32ReadOnlyFile, now were going thru everything again but this time were going to use CStdioFile.
If it now fails, the file doesn't exist.